+++ src/lib/services/types/verification/evm.ts
@@ -1,7 +1,7 @@
 import { JsonFragment } from "ethers";
 import { zHexAddr20, zUtcDate } from "lib/types";
-import { snakeToCamel } from "lib/utils";
-import { z } from "zod";
+import { isHex20Bytes, snakeToCamel } from "lib/utils";
+import { z, ZodIssueCode } from "zod";
 
 export enum EvmProgrammingLanguage {
   Solidity = "Solidity",
@@ -66,6 +66,123 @@ const zEvmOptimizer = z.object({
   runs: z.number(),
 });
 
+export enum EvmVerifyOptions {
+  VyperUploadFile = "vyper-upload-file",
+  VyperContractCode = "vyper-contract-code",
+  VyperJsonInput = "vyper-json-input",
+  SolidityUploadFiles = "solidity-upload-files",
+  SolidityContractCode = "solidity-contract-code",
+  SolidityJsonInput = "solidity-json-input",
+  SolidityHardhat = "solidity-hardhat",
+  SolidityFoundry = "solidity-foundry",
+}
+
+// MARK - Shared Components
+const zOptimizerConfig = z.object({
+  enabled: z.boolean(),
+  runs: z.string(),
+});
+export type OptimizerConfig = z.infer<typeof zOptimizerConfig>;
+
+const zConstructorArgs = z.object({
+  enabled: z.boolean(),
+  value: z.string(),
+});
+
+const zEvmTargetVersion = z.string();
+
+const zContractLibrary = z.object({
+  name: z.string(),
+  address: zHexAddr20,
+});
+
+const zContractLibraries = z.object({
+  enabled: z.boolean(),
+  value: z.array(zContractLibrary),
+});
+
+// MARK - Solidity
+const zEvmContractVerifySolidityOptionUploadFilesForm = z.object({
+  files: z.array(
+    z.object({
+      file: z.instanceof(File),
+    })
+  ),
+  constructorArgs: zConstructorArgs,
+  optimizerConfig: zOptimizerConfig,
+  evmVersion: zEvmTargetVersion,
+  contractLibraries: zContractLibraries,
+});
+
+const zEvmContractVerifySolidityOptionContractCodeForm = z.object({
+  contractCode: z.string(),
+  constructorArgs: zConstructorArgs,
+  evmVersion: zEvmTargetVersion,
+  optimizerConfig: zOptimizerConfig,
+  contractLibraries: zContractLibraries,
+});
+
+const zEvmContractVerifySolidityOptionJsonInputForm = z.object({
+  jsonFile: z.instanceof(File).optional(),
+  constructorArgs: zConstructorArgs,
+});
+
+// MARK - Vyper
+const zEvmContractVerifyVyperOptionUploadFileForm = z.object({
+  file: z.instanceof(File).optional(),
+  constructorArgs: zConstructorArgs,
+  evmVersion: zEvmTargetVersion,
+});
+
+const zEvmContractVerifyVyperOptionContractCodeForm = z.object({
+  contractName: z.string(),
+  contractCode: z.string(),
+  constructorArgs: zConstructorArgs,
+  evmVersion: zEvmTargetVersion,
+});
+
+const zEvmContractVerifyVyperOptionJsonInputForm = z.object({
+  jsonFile: z.instanceof(File).optional(),
+  constructorArgs: zConstructorArgs,
+});
+
+// MARK - Base
+const zEvmContractVerifyBase = z.object({
+  // TODO: refactor later
+  contractAddress: zHexAddr20.superRefine((val, ctx) => {
+    if (val === "")
+      ctx.addIssue({
+        code: ZodIssueCode.custom,
+        message: " ",
+      });
+
+    if (!isHex20Bytes(val))
+      ctx.addIssue({
+        code: ZodIssueCode.custom,
+        message: "Invalid address",
+      });
+  }),
+  compilerVersion: z.string(),
+  licenseType: z.nativeEnum(EvmVerifyLicenseType).optional(),
+  language: z.nativeEnum(EvmProgrammingLanguage).optional(),
+  option: z.nativeEnum(EvmVerifyOptions).optional(),
+});
+
+export const zEvmContractVerifyForm = zEvmContractVerifyBase.merge(
+  z.object({
+    verifyForm: z.object({
+      solidityUploadFiles: zEvmContractVerifySolidityOptionUploadFilesForm,
+      solidityContractCode: zEvmContractVerifySolidityOptionContractCodeForm,
+      solidityJsonInput: zEvmContractVerifySolidityOptionJsonInputForm,
+      vyperUploadFile: zEvmContractVerifyVyperOptionUploadFileForm,
+      vyperContractCode: zEvmContractVerifyVyperOptionContractCodeForm,
+      vyperJsonInput: zEvmContractVerifyVyperOptionJsonInputForm,
+    }),
+  })
+);
+export type EvmContractVerifyForm = z.infer<typeof zEvmContractVerifyForm>;
+
+// MARK - EvmVerifyInfo
 export const zEvmVerifyInfo = z
   .object({
     guid: z.string().uuid(),
@@ -77,23 +194,43 @@ export const zEvmVerifyInfo = z
     contract_path: z.string().optional(),
     compiler_version: z.string(),
     evm_version: z.string(),
-    optimizer: z.string(),
+    optimizer: z.string().optional(),
     constructor_arguments: z.string(),
     abi: z.string().optional(),
     bytecode_type: z.string(),
     verified_timestamp: zUtcDate.optional(),
     submitted_timestamp: zUtcDate,
     error_message: z.string().optional(),
-    settings: z.string(),
+    settings: z.string().default("{}"),
     source_files: z.array(zEvmVerifyInfoSourceFile),
   })
   .transform(({ abi, optimizer, ...rest }) => ({
     ...snakeToCamel(rest),
     abi: abi ? (JSON.parse(abi) as JsonFragment[]) : [],
-    optimizer: zEvmOptimizer.parse(JSON.parse(optimizer)),
+    optimizer: optimizer
+      ? zEvmOptimizer.parse(JSON.parse(optimizer))
+      : undefined,
     isVerified: !!rest.verified_timestamp,
     libraries: zEvmVerifyInfoLibraries.parse(
       JSON.parse(rest.settings).libraries
     ),
   }));
 export type EvmVerifyInfo = z.infer<typeof zEvmVerifyInfo>;
+
+// MARK - SubmitEvmVerify
+type SubmitEvmVerifyBaseArgs = {
+  verifierUrl: string;
+  contractAddress: string;
+  chainId: string;
+  compilerVersion: string;
+  licenseType: EvmVerifyLicenseType;
+};
+
+export type SubmitEvmVerifyFlattenArgs = SubmitEvmVerifyBaseArgs &
+  EvmContractVerifyForm["verifyForm"]["solidityContractCode"];
+
+export interface SubmitEvmVerifyArgs
+  extends Omit<SubmitEvmVerifyBaseArgs, "verifierUrl"> {
+  option: EvmVerifyOptions;
+  verifyForm: EvmContractVerifyForm["verifyForm"];
+}