+++ src/lib/services/verification/evm/api.ts
@@ -1,7 +1,15 @@
 import axios from "axios";
 import { CELATONE_VERIFICATION_API } from "env";
-import { zEvmVerifyConfig, zEvmVerifyInfo } from "lib/services/types";
+import {
+  EvmVerifyOptions,
+  SubmitEvmVerifyArgs,
+  SubmitEvmVerifyFlattenArgs,
+  zEvmVerifyConfig,
+  zEvmVerifyInfo,
+} from "lib/services/types";
 import { HexAddr20 } from "lib/types";
+import { getVerifierUrl } from "./utils";
+import { isHex20Bytes } from "lib/utils";
 
 export const getEvmVerifyConfig = async () =>
   axios
@@ -11,8 +19,8 @@ export const getEvmVerifyConfig = async () =>
 export const getEvmVerifyInfo = async (
   chainId: string,
   contractAddress: HexAddr20
-) => {
-  return axios
+) =>
+  axios
     .get(`${CELATONE_VERIFICATION_API}/evm/verification/info`, {
       params: {
         chain_id: chainId,
@@ -20,4 +28,93 @@ export const getEvmVerifyInfo = async (
       },
     })
     .then(({ data }) => zEvmVerifyInfo.parse(data));
+
+// === EVM Contract Verification Submission ===
+// Handles HTTP POST requests for different verification options.
+// Prepares data for Solidity & Vyper verification.
+// ============================================
+
+const submitEvmVerifyFlatten = async ({
+  verifierUrl,
+  contractAddress,
+  chainId,
+  compilerVersion,
+  licenseType,
+  contractCode,
+  optimizerConfig,
+  constructorArgs,
+  evmVersion,
+  contractLibraries,
+}: SubmitEvmVerifyFlattenArgs) => {
+  const settings = {
+    evmVersion: evmVersion === "default" ? "cancun" : evmVersion,
+    optimizer: {
+      enabled: optimizerConfig.enabled,
+      runs: Number(optimizerConfig.runs),
+    },
+    libraries: contractLibraries.value.reduce((acc, library) => {
+      if (
+        library.name.trim().length === 0 ||
+        !isHex20Bytes(library.address.trim())
+      )
+        return acc;
+
+      Object.assign(acc, {
+        [`${library.name}.sol`]: {
+          [library.name]: library.address,
+        },
+      });
+      return acc;
+    }, {}),
+    metadata: {
+      bytecodeHash: "none",
+    },
+    outputSelection: {
+      "*": {
+        "": ["ast"],
+        "*": ["abi", "evm.bytecode", "evm.methodIdentifiers", "metadata"],
+      },
+    },
+  };
+
+  return axios.post(
+    verifierUrl,
+    {
+      license: licenseType,
+      bytecode_type: "CREATION_INPUT",
+      compiler_version: compilerVersion,
+      constructor_arguments: constructorArgs.value,
+      metadata: {
+        chain_id: chainId,
+        contract_address: contractAddress,
+      },
+      source_code: contractCode,
+      settings: JSON.stringify(settings),
+    },
+    {
+      headers: {
+        "Content-Type": "application/json",
+      },
+    }
+  );
+};
+
+export const submitEvmVerify = async ({
+  option,
+  verifyForm,
+  ...rest
+}: SubmitEvmVerifyArgs) => {
+  const verifierUrl = getVerifierUrl(option);
+
+  switch (option) {
+    case EvmVerifyOptions.SolidityContractCode:
+      return submitEvmVerifyFlatten({
+        verifierUrl,
+        ...rest,
+        ...verifyForm.solidityContractCode,
+      });
+    // TODO: Implement other options
+    default:
+      throw new Error(`Unsupported option: ${option}`);
+  }
 };