+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1220](https://github.com/alleslabs/celatone-frontend/pull/1220) Add EVM contract verification for Solidity with contract code (flattened method)
 - [#1219](https://github.com/alleslabs/celatone-frontend/pull/1219) Add tooltip to full editor sidebar
 - [#1218](https://github.com/alleslabs/celatone-frontend/pull/1218) Add EVM library contract
 - [#1217](https://github.com/alleslabs/celatone-frontend/pull/1217) Add evm contract method shortcuts on the overview section